requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==4.2.13
django-filter==24.2
django-split-settings==1.2.0
djangorestframework==3.15.1
geographiclib==2.0
geopy==2.4.1
idna==3.7
mysqlclient==2.2.4
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.1
urllib3==2.2.1
